+++ internal/extsvc/github/v4.go
@@ -90,6 +90,9 @@ func NewV4Client(urn string, apiURL *url.URL, a auth.Authenticator, cli httpcli.
 
 	rl := ratelimit.DefaultRegistry.Get(urn)
 	rlm := ratelimit.DefaultMonitorRegistry.GetOrSet(apiURL.String(), tokenHash, "graphql", &ratelimit.Monitor{HeaderPrefix: "X-"})
+	if rlm.Collector() == nil {
+		addMetricsCollector(rlm, "graphql", a.Hash())
+	}
 
 	return &V4Client{
 		log:              log.Scoped("github.v4", "github v4 client"),
